requirements.txt
djoser
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0
django-cors-headers==3.2.0
djangorestframework==3.11.0
google-auth==1.8.2
google-auth-oauthlib==0.4.1
googleads==22.0.0
httplib2==0.14.0
idna==2.8
isodate==0.6.0
lxml==4.4.2
numpy==1.17.4
oauth2client==4.1.3
oauthlib==3.1.0
pandas==0.25.3
Pillow==6.2.1
praat-parselmouth==0.3.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scipy==1.3.3
six==1.13.0
sqlparse==0.3.0
stopit==1.1.2
suds-jurko==0.6
urllib3==1.25.7
xmltodict==0.12.0
zeep==3.4.0
